refactor sqgen-postgres so config has pointer to sql.DB
