+++ sqgen/postgres/db.go
@@ -1,33 +1,9 @@
 package postgres
 
 import (
-	"database/sql"
-	"fmt"
 	"strconv"
 	"strings"
-
-	_ "github.com/lib/pq"
 )
-
-func openAndPing(database string) (*sql.DB, error) {
-	db, err := sql.Open("postgres", database)
-
-	if err != nil {
-		return nil, err
-	}
-
-	err = db.Ping()
-
-	if err != nil {
-		return nil, fmt.Errorf(
-			"Could not ping the database, is the database reachable via %s? %w",
-			database,
-			err,
-		)
-	}
-
-	return db, nil
-}
 
 // replacePlaceholders will replace question mark placeholders with dollar
 // placeholders e.g. ?, ?, ? -> $1, $2, $3 etc
